fix: Return if there is no request object instead of runtime error
The authenticate method may be called without a request object such as
when it is provided a username and password as keyword arguments. This
is useful for other authentication backends but token auth always
requires a request object. If there is none the method returns early
preventing an exception later on